// src/app/interfaces/ISupervisor.ts
export interface ISupervisorInsert {
  CPD: number;
  Odd_hours: number;
  rewards_earning: number;
  content_creation: number;
  event_participate: number;
  survey_score: string;
  LP_submission: string;
  report_to_student: string;
  student_assign_feedback: string;
  result_submission_to_FD: string;

}

export interface ISupervisorSingle {

}